Simplified import statements; updated SCSS structure and code styling
This commit addresses a code cleanup in various java entities, and modified the import statements for better readability and efficiency. Simplified TrainingPlan model.
In the style.scss file, the relative paths for the imported styles were simplified to avoid unnecessary complicated path

The spring-boot-starter parent version in the pom.xml file was updated for better performance and to receive the latest bug fixes. The same applies to other dependencies.

Gradually transitioning the codebase towards a cleaner, easier to manage structure. This commit also removes commented out code and reformat code for improved readability.